priority_classification/requirements.txt
fastapi
uvicorn

requests
python-dotenv

pydantic

pytest
